fixed variable inputs

+++ ParticleHLF.py
@@ -7,7 +7,7 @@ class ParticleHLF(Dataset):
         self.len = data_y.shape[0]
         self.data_particles = torch.from_numpy(data_particles).float()
         self.data_hlf = torch.from_numpy(data_hlf).float()
-        self.data_y = torch.from_numpy(data_y).float()
+        self.data_y = torch.from_numpy(data_y).long()
         
     def __len__(self):
         return self.len
